requirements.txt
dash==2.8.1
dash_bootstrap_components==1.2.1
numpy==1.24.4
pandas==1.3.5
plotly==5.5.0
python_dateutil==2.9.0.post0
yfinance==0.2.51
gunicorn
dash-tools
